Adds covariant return types for AstExp subclasses

+++ src/test/java/nl/jqno/annotationscript/language/ast/AstFloatTest.java
@@ -8,7 +8,7 @@ public class AstFloatTest {
     @Test
     public void value() {
         var sut = new AstFloat(1.2);
-        assertEquals(1.2, (double)sut.value(), 0.0001);
+        assertEquals(1.2, sut.value(), 0.0001);
         assertEquals(1.2, sut.asFloat(), 0.0001);
     }
